Move implementation to task exit
